% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_seir.R
\name{gener.seir}
\alias{gener.seir}
\title{Generates a course of an SEIR epidemic.}
\usage{
gener.seir(lambda, mu, nu, n.pop = 100, m = 1, max.infections = n.pop)
}
\arguments{
\item{lambda}{the rate of the Poisson process, which generates the points
when the infection can occur.}

\item{mu}{the rate of the exponential distribution of the infectious periods.}

\item{nu}{the rate of the exponential distribution of the latent periods.}

\item{n.pop}{the total population size excluding the initial infectious
individuals.}

\item{m}{the number of initial infectious individuals at the beginning of the
epidemic.}

\item{max.infections}{how many infections can occur before the epidemic
stops. Use especially for the ABC algorithm  when the population is large
to reduce the amount of time spent on generating a single epidemic, which
is too different from the observed one anyway. The value of maximal
infections must be set manually regarding the data.}
}
\value{
A list of three numerical vectors \code{E}, \code{I} and \code{R}
    and a logical value \code{stopped}. Initial infectious individuals are
    included in the first \code{m} positions. Hence all three vectors are of
    length \code{m} + \code{n.pop}.
    \itemize{
    \item \code{I} the infection times for each individual
    \item \code{E} the exposition times for each individual
    \item \code{R} the recovery times for each individual
    \item \code{stopped} logical value, if the epidemic was stopped after
    exceeding the \code{max.infections} number
    }
}
\description{
Generates a course of an SEIR epidemic. Returns vector of
infection, exposition and removal times.
}
\examples{
gener.seir(n.pop = 1000, m = 5, lambda = 0.6, mu = 0.3, nu = 1)
}
