% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_sveird.R
\name{gener.sveird}
\alias{gener.sveird}
\title{Generates a course of an SVEIRD epidemic.}
\usage{
gener.sveird(
  lambda,
  mu,
  nu,
  psi,
  phi,
  omega,
  kappa,
  max.duration,
  s0 = 100,
  i0 = 1,
  v0 = 0,
  r0 = 0,
  max.infections = 10000,
  all.comps = TRUE
)
}
\arguments{
\item{lambda}{the rate of the Poisson process, which generates the points
when the infection can occur.}

\item{mu}{the rate of the exponential distribution of the time from
the start of the infectiousness to the time of recovery.}

\item{nu}{the rate of the exponential distribution of the latent periods.}

\item{psi}{the rate of the exponential distribution of the time before
the first vaccine dose.}

\item{omega}{the rate of the exponential distribution of the time, after
which the immunity gained by the recovery from the disease fades out
and the individual becomes susceptible again.}

\item{kappa}{the rate of the exponential distribution of the time from
the start of the infectiousness to the time of death.}

\item{max.duration}{the maximum duration of the epidemic, which must be set
manually, since in this epidemic model does not guarantee that the
algorithm ends in a reasonable time.}

\item{s0}{the initial susceptible population size.}

\item{i0}{the number of initial infectious individuals.}

\item{v0}{the number of initial vaccinated individuals.}

\item{r0}{the number of initial recovered individuals.}

\item{max.infections}{how many infections can occur before the epidemic
stops. Use especially for the ABC algorithm  when the population is large
to reduce the amount of time spent on generating a single epidemic, which
is too different from the observed one anyway. The value of maximal
infections must be set manually regarding the data.}

\item{all.comps}{indicator, whether the outcome should include entry times
of all compartments. If \code{FALSE}, only the 'E', 'I', 'R' and 'D'
compartments are included.}
}
\value{
A list of three lists \code{I}, \code{V}, \code{R}, one vector 
    \code{D}, one numeric value \code{duration} one character vector and a 
    logical value \code{stopped}. If \code{all.comps = TRUE}, additional 
    lists \code{S} and \code{V} are included. Initial infectious individuals
    are included in the first \code{i0} positions of the lists/vector. Hence 
    all three/five lists and both vectors are of length 
    \code{i0} + \code{s0}. The elements in the six lists are in the 
    decreasing order for the sake of convenience of the algorithm.
    \itemize{
    \item \code{S} the renewed susceptibility times for each individual
    \item \code{E} the exposition times for each individual
    \item \code{I} the infection times for each individual
    \item \code{V} the vaccination times for each individual
    \item \code{R} the recovery times for each individual
    \item \code{D} the death times for each individual
    \item \code{stopped} logical value, if the epidemic was stopped after
    exceeding the \code{max.infections} number
    \item \code{duration} the duration of the epidemic, which is less or 
    equal to the \code{max.duration} number.
    }
}
\description{
Generates a course of an SVEIRD epidemic. Returns lists of
infection, exposition, removal, death, vaccination and renewed 
susceptibility times.
}
\examples{
gener.sveird(lambda = 2, mu = 0.5, nu = 0.25, psi = 0.2, omega = 0.05,
             phi = 0.1, max.duration = 100, kappa = 0.1, s0 = 1000,
             i0 = 1, v0 = 5, r0 = 10, all.comps = TRUE)
}
