% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_proc.R
\name{poiss.proc.points}
\alias{poiss.proc.points}
\title{Generate jump times of a Poisson process}
\usage{
poiss.proc.points(lambda, duration)
}
\arguments{
\item{lambda}{the rate of the Poisson process.}

\item{duration}{a numeric value specifying the maximal duration of the
process.}
}
\value{
a numeric vector of jump times of the Poisson process. If no event
    occurred until the \code{duration}, a vector of length 0 is returned.
}
\description{
The function generates a sequence of jump times of a Poisson
process of rate \code{lambda}.
}
\examples{
poiss.proc.points(lambda = 1, duration = 2.5)
}
