% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_ABC.R
\name{seir.ABC}
\alias{seir.ABC}
\title{ABC for an SEIR epidemic model}
\usage{
seir.ABC(
  I.obs,
  n.pop,
  m,
  max.infections = n.pop,
  times = 100,
  max.times = 10000,
  tolerance = 100,
  transf = NULL,
  kern = NULL,
  dist.metr = "euclidean",
  prior = list(lambda.samp = runif, mu.samp = runif, nu.samp = runif, lambda.dens = NULL,
    mu.dens = NULL, nu.dens = NULL),
  importance = list(lambda.samp = NULL, mu.samp = NULL, nu.samp = NULL, lambda.dens =
    NULL, mu.dens = NULL, nu.dens = NULL),
  prior.params = list(...),
  importance.params = list(...)
)
}
\arguments{
\item{I.obs}{A positive integer valued vector of daily case counts. Days
with zero cases must be included.}

\item{n.pop}{A positive integer specifying the total population size
excluding the initial infectious individuals.}

\item{m}{A positive integer specifying the number of infectious individuals
at the beginning of the epidemic.}

\item{max.infections}{A positive integer specifying the maximum number of 
infections to occur in a single epidemic. If the epidemic tends to generate
more infections, it stops immediately and the proposed particle in the ABC 
algorithm is rejected. Use especially for the ABC algorithm when the 
population is large to reduce the amount of time spent on generating a
single epidemic. The value of maximal infections must be set manually 
regarding the data I.obs.}

\item{times}{A desired size of the sample from the approximate posterior
distribution}

\item{max.times}{A maximal number of iterations}

\item{tolerance}{A positive number specifying the tolerance for accepting 
proposed particles.}

\item{transf}{A function transforming the case count. Must be specified
with one parameters - a case counts vector. The population size and the
number of initial infectiouscindividuals can be used too, since these 
variables exist within the 'sir.ABC' function. Their names are 'n.pop' and
'm' respectively.}

\item{kern}{A kernel non-negative function, symmetric
around zero, with maximum at zero. Must be specified without any additional 
parameters, i.e. Must be a function of only the value at which we wish to 
evaluate the kernel.}

\item{dist.metr}{Specification of the Mahalanobis distance metric between 
the observed and simulated summary statistic. Must be specified as an 
inverse of a positive definite matrix of the distance metric or the string 
"euclidean". In the latter case, standard euclidean distance is used. 
Use the matrix option only if the dimension of the summary statistics is 
known beforehands!}

\item{prior}{A named list specifying the prior distribution of 'lambda' and 
'mu'. Both distributions are considered independent! Therefore the joint
distribution is a product of the marginals. Must be either of the form:
list(lambda.samp = prior distribution of 'lambda',
mu.samp = prior distribution of 'mu', nu.samp = prior distribution of 'nu')
for the Rejection sampling variant, or of the form: 
list(lambda.dens = prior density of 'lambda', 
mu.dens = prior density of 'mu', nu.dens = prior density of 'nu')
for the Importance sampling variant. In case of the Importance sampling, 
the sampler and the density functions accept the same parameters passed 
from 'prior.params'.}

\item{importance}{A named list specifying the importance distribution of 
'lambda', 'mu' and 'nu'. Must be of the form: 
list(lambda.samp = importance sampler of 'lambda',
mu.samp = importance sampler of 'mu', lambda.dens = importance density of 'lambda',
mu.dens = importance density of 'mu'). The sampler and the density functions
accept the same parameters passed from 'prior.params'.}

\item{prior.params}{A named list specifying parameters of the prior 
distribution of 'lambda' and 'mu'. These prior parameters are passed to 
the functions specified in the parameter 'prior'. Must be of the form:
list(lambda = c(params for the prior sampler and density of lambda),
mu = c(params for the prior sampler and density of mu)).}

\item{importance.params}{A named list specifying parameters of the 
importance distribution of 'lambda' and 'mu'. These prior parameters are 
passed to the functions specified in the parameter 'importance'. Must be of 
the form: 
list(lambda = c(params for the importance sampler and density of lambda),
mu = c(params for the importance sampler and density of mu)).}
}
\value{
A list consisting of five to seven elements: a numerical matrix containing 
accepted particles, a numerical matrix containing accepted particles
adjusted by the linear regression, a list of summary statistics, total 
number of iterations and number of iterations, which resulted in accepting
the proposed particle. In case of importance sampling, there is an 
additional numerical matrix of importance weights. If a transformation
is used, then list of non-transformed case counts is returned too.
}
\description{
Performs the classic Approximate Bayesian computation algorithm
for the SEIR stochastic model.
}
\examples{

transf <- function (x) {
 log(x + 1)
}
gse <- gener.sir(n.pop = 100, m = 5, lambda = 1.5, mu = 0.5)
I <- as.vector(table(floor(gse$I)))

# The Importance sampling variant without kernel, with the prior 
# distribution
Unif(0.5, 2), Unif(0.1, 1.1) for lambda and mu respectively and the 
importance distribution Gamma(1.5, 1) for both lambda and mu.
sir.ABC(
I, n.pop = 100, m = 5, times = 1e2, max.times = 1e3, tolerance = 50, 
prior = c(lambda.samp = runif, mu.samp = runif, 
          lambda.dens = dunif, mu.dens = dunif),
prior.params = list(lambda = c(min = 0.5, max = 2), 
                    mu = c(min = 0.1, max = 1.1)), 
importance = list(lambda.samp = rgamma, mu.samp = rgamma,
                  lambda.dens = dgamma, mu.dens = dgamma),
importance.params = list(lambda = c(shape = 1.5, rate = 1),
                         mu = c(shape = 1.5, rate = 1))
)

The Rejection sampling variant with the standardised normal kernel, with 
the prior distribution Unif(0.5, 2), Unif(0.1, 1.1) for lambda and mu 
respectively and the importance distribution Gamma(1.5, 1) for both lambda 
and mu. A transformation 'transf' is used.
sir.ABC(
I, n.pop = 100, m = 5, times = 1e2, max.times = 1e3, tolerance = 3, 
kern = dunif, transf = transf,
prior = c(lambda.samp = runif, mu.samp = runif),
prior.params = list(lambda = c(min = 0.5, max = 2), 
                    mu = c(min = 0.1, max = 1.1))
)
}
